+++ pdoc/__init__.py
@@ -340,7 +340,7 @@ We would be happy to accept contributions for Markdown and PDF.
 '''
 from __future__ import annotations
 
-__version__ = "6.6.0"  # this is read from setup.py
+__version__ = "7.0.0"  # this is read from setup.py
 
 import io
 import traceback